requirements.txt
# Visual RAG-Lite Framework Requirements

# Core dependencies
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0

# OCR and Image Processing
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Embeddings and Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA is available
clip>=1.0

# Evaluation and Metrics
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
Levenshtein>=0.21.0

# Data Handling
pandas>=2.0.0
datasets>=2.14.0
tqdm>=4.65.0

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.15.0

# Utilities
PyYAML>=6.0
jsonlines>=3.1.0
argparse

# Development and Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
